hardcode minimum for partial liquidations (#269)
Add a minimum value, under which there's no constraint that a partial
amount of the collateral is liquidated. This is useful to prevent cases
where accounts are slowly liquidated to a point where the remaining
balances are too small and get rounded to zero when attempting to repay
via liquidation.